2nd error print, printed the first error
No error pointer was passed in before. That means if opening a Realm with configuration retrieved via `DBRealmModel1` would have caused an error, you wouldn't have seen any 2nd or 3rd error print on the console. But instead it succeeds, as expected, and prints the error which happened when trying to migrate the default.realm with a to low schema version.